Refractor all the code
